Move and fix some issues with tmpro effect
