+++ Assets/Scenes/SampleScene.unity
@@ -7827,6 +7827,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c2c09953ccd10cb4d8a290d4da849f29, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  effects: 02000000
   refreshRate: 0.1
   textGUI: {fileID: 411621256}
   pulseColorA:
@@ -7837,7 +7838,25 @@ MonoBehaviour:
     rgba: 4294967295
   pulseAnimation:
     serializedVersion: 2
-    m_Curve: []
+    m_Curve:
+    - serializedVersion: 3
+      time: 0
+      value: 0
+      inSlope: 0
+      outSlope: 0
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0
+      outWeight: 0
+    - serializedVersion: 3
+      time: 1
+      value: 1
+      inSlope: 0
+      outSlope: 0
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0
+      outWeight: 0
     m_PreInfinity: 2
     m_PostInfinity: 2
     m_RotationOrder: 4
