--- Assets/Scripts/Utils.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b847a25da9e1a3f449a745cae82219c1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
